Remove duplicate stats aggregator from V2 query metrics
